created stories folder in themeContext
