+++ src/js/contexts/ThemeContext/stories/ThemeContextExtend.js
@@ -5,24 +5,6 @@ import { grommet } from 'grommet/themes';
 import { deepMerge } from 'grommet/utils';
 import { Box, Text, ThemeContext, Grommet, Select } from 'grommet';
 
-const ExternalComponentWithTheme = () => (
-  <Grommet theme={grommet}>
-    <Box background="neutral-3">
-      <Text color="light-1">This is a grommet component</Text>
-    </Box>
-    <ThemeContext.Consumer>
-      {theme => (
-        <div style={{ backgroundColor: theme.global.colors['neutral-3'] }}>
-          <p style={{ color: theme.global.colors['light-1'] }}>
-            This component is leveraging the grommet theme capabilities although
-            it is not a grommet component
-          </p>
-        </div>
-      )}
-    </ThemeContext.Consumer>
-  </Grommet>
-);
-
 const customTheme = deepMerge(grommet, {
   global: {
     focus: {
@@ -84,6 +66,6 @@ const GlobalThemeWithThemeContext = () => {
   );
 };
 
-storiesOf('Theme', module)
-  .add('External Components', () => <ExternalComponentWithTheme />)
-  .add('ThemeContext.Extend', () => <GlobalThemeWithThemeContext />);
+storiesOf('Theme', module).add('ThemeContext.Extend', () => (
+  <GlobalThemeWithThemeContext />
+));
